requirements.txt
transformers
torch
pandas
requests # Ensure this line exists
langchain
huggingface_hub
python-dotenv
scikit-learn # Needed for tabular simulation
numpy # Needed for tabular simulation
Pillow # For image processing (Vision, Multimodal)
decord # For video processing (Multimodal)
pyarrow # Needed for Parquet format support
pythainlp # Added for potential Thai language processing tasks
webdataset # Needed for WebDataset format support

# Optional Audio Libraries (uncomment if needed for specific audio tasks)
# pydub
# librosa
# soundfile
